requirements.txt
cdsapi
earthkit-data
earthkit-plots
cartopy
numpy
matplotlib
netCDF4
cfgrib
eccodes